refactor(cli): move pprof dump logic into src/pprof_dump.rs
Keep main.rs thin: the SIGUSR2 profiler-dump machinery (guard, signal
handler, watcher thread, report writer) now lives in a bin-private
pprof_dump module behind a small start() -> Watcher::shutdown() API.
